Refactored most Handlers into the eventHandlers Package.
Removed some commented code

Replaced pythagorize method with Math.hypot(a,b)
